frontend refactor: moved getHostname function from EventRepository.kt to UserRepository.kt to better follow the app architecture plan
